[gui] move visualizer classes into gui.visualizers module
