+++ src/pymor/gui/qt.py
@@ -14,20 +14,16 @@ import numpy as np
 import multiprocessing
 import os
 import signal
-import time
 
 from pymor.core.config import config
 from pymor.core.defaults import defaults
-from pymor.core.interfaces import BasicInterface
 from pymor.core.logger import getLogger
 from pymor.core.exceptions import QtMissing
-from pymor.grids.oned import OnedGrid
-from pymor.grids.referenceelements import triangle, square
 from pymor.gui.gl import GLPatchWidget, ColorBarWidget
 from pymor.gui.matplotlib import Matplotlib1DWidget, MatplotlibPatchWidget
 from pymor.tools.vtkio import write_vtk
 from pymor.vectorarrays.interfaces import VectorArrayInterface
-from pymor.vectorarrays.numpy import NumpyVectorArray, NumpyVectorSpace
+from pymor.vectorarrays.numpy import NumpyVectorSpace
 
 if config.HAVE_QT:
     from Qt.QtWidgets import (QWidget, QVBoxLayout, QHBoxLayout, QGridLayout, QSlider, QApplication, QLCDNumber,
@@ -434,129 +430,3 @@ def visualize_matplotlib_1d(grid, U, codim=1, title=None, legend=None, separate_
             self.grid = grid
 
     _launch_qt_app(lambda: MainWindow(grid, U, codim, title=title, legend=legend, separate_plots=separate_plots), block)
-
-
-class PatchVisualizer(BasicInterface):
-    """Visualize scalar data associated to a two-dimensional |Grid| as a patch plot.
-
-    The grid's |ReferenceElement| must be the triangle or square. The data can either
-    be attached to the faces or vertices of the grid.
-
-    Parameters
-    ----------
-    grid
-        The underlying |Grid|.
-    bounding_box
-        A bounding box in which the grid is contained.
-    codim
-        The codimension of the entities the data in `U` is attached to (either 0 or 2).
-    backend
-        Plot backend to use ('gl' or 'matplotlib').
-    block
-        If `True`, block execution until the plot window is closed.
-    """
-
-    def __init__(self, grid, bounding_box=([0, 0], [1, 1]), codim=2, backend=None, block=False):
-        assert grid.reference_element in (triangle, square)
-        assert grid.dim == 2
-        assert codim in (0, 2)
-        self.grid = grid
-        self.bounding_box = bounding_box
-        self.codim = codim
-        self.backend = backend
-        self.block = block
-
-    def visualize(self, U, discretization, title=None, legend=None, separate_colorbars=False,
-                  rescale_colorbars=False, block=None, filename=None, columns=2):
-        """Visualize the provided data.
-
-        Parameters
-        ----------
-        U
-            |VectorArray| of the data to visualize. If `len(U) > 1`, the data is visualized
-            as a time series of plots. Alternatively, a tuple of |VectorArrays| can be
-            provided, in which case a subplot is created for each entry of the tuple. The
-            lengths of all arrays have to agree.
-        discretization
-            Filled in :meth:`pymor.discretizations.DiscretizationBase.visualize` (ignored).
-        title
-            Title of the plot.
-        legend
-            Description of the data that is plotted. Most useful if `U` is a tuple in which
-            case `legend` has to be a tuple of strings of the same length.
-        separate_colorbars
-            If `True`, use separate colorbars for each subplot.
-        rescale_colorbars
-            If `True`, rescale colorbars to data in each frame.
-        block
-            If `True`, block execution until the plot window is closed. If `None`, use the
-            default provided during instantiation.
-        filename
-            If specified, write the data to a VTK-file using
-            :func:`pymor.tools.vtkio.write_vtk` instead of displaying it.
-        columns
-            The number of columns in the visualizer GUI in case multiple plots are displayed
-            at the same time.
-        """
-        assert isinstance(U, VectorArrayInterface) and hasattr(U, 'data') \
-            or (isinstance(U, tuple) and all(isinstance(u, VectorArrayInterface) and hasattr(u, 'data') for u in U)
-                and all(len(u) == len(U[0]) for u in U))
-        if filename:
-            if not isinstance(U, tuple):
-                write_vtk(self.grid, U, filename, codim=self.codim)
-            else:
-                for i, u in enumerate(U):
-                    write_vtk(self.grid, u, '{}-{}'.format(filename, i), codim=self.codim)
-        else:
-            block = self.block if block is None else block
-            visualize_patch(self.grid, U, bounding_box=self.bounding_box, codim=self.codim, title=title,
-                            legend=legend, separate_colorbars=separate_colorbars, rescale_colorbars=rescale_colorbars,
-                            backend=self.backend, block=block, columns=columns)
-
-
-class Matplotlib1DVisualizer(BasicInterface):
-    """Visualize scalar data associated to a one-dimensional |Grid| as a plot.
-
-    The grid's |ReferenceElement| must be the line. The data can either
-    be attached to the subintervals or vertices of the grid.
-
-    Parameters
-    ----------
-    grid
-        The underlying |Grid|.
-    codim
-        The codimension of the entities the data in `U` is attached to (either 0 or 1).
-    block
-        If `True`, block execution until the plot window is closed.
-    """
-
-    def __init__(self, grid, codim=1, block=False):
-        assert isinstance(grid, OnedGrid)
-        assert codim in (0, 1)
-        self.grid = grid
-        self.codim = codim
-        self.block = block
-
-    def visualize(self, U, discretization, title=None, legend=None, block=None):
-        """Visualize the provided data.
-
-        Parameters
-        ----------
-        U
-            |VectorArray| of the data to visualize. If `len(U) > 1`, the data is visualized
-            as a time series of plots. Alternatively, a tuple of |VectorArrays| can be
-            provided, in which case several plots are made into the same axes. The
-            lengths of all arrays have to agree.
-        discretization
-            Filled in by :meth:`pymor.discretizations.DiscretizationBase.visualize` (ignored).
-        title
-            Title of the plot.
-        legend
-            Description of the data that is plotted. Most useful if `U` is a tuple in which
-            case `legend` has to be a tuple of strings of the same length.
-        block
-            If `True`, block execution until the plot window is closed. If `None`, use the
-            default provided during instantiation.
-        """
-        block = self.block if block is None else block
-        visualize_matplotlib_1d(self.grid, U, codim=self.codim, title=title, legend=legend, block=block)
